Testing with different uv textures to get a smooth outline.

+++ Project/Classes/Lightning.cpp
@@ -71,7 +71,7 @@ void Lightning::Render() {
 
 	//XMViewSpace = XMMatrixLookAtLH(camPos2, camLook2, camUp2);
 	XMViewSpace = XMMatrixLookAtLH(XMCamPos, XMCamLook, XMCamUp);
-	Projection = XMMatrixPerspectiveFovLH(3.14*0.45, m_graphicsManager->getWindowWidth() / m_graphicsManager->getWindowHeight(), 0.01f, 20.0f); //	FLOAT FovAngleY, FLOAT AspectRatio, FLOAT NearZ, FLOAT FarZ
+	Projection = XMMatrixPerspectiveFovLH(3.14f*0.45f, (float)m_graphicsManager->getWindowWidth() / (float)m_graphicsManager->getWindowHeight(), 0.01f, 20.0f); //	FLOAT FovAngleY, FLOAT AspectRatio, FLOAT NearZ, FLOAT FarZ
 	//Projection = XMMatrixOrthographicLH(20, 20, 0.5f, 20.0f);
 	World = XMMatrixIdentity();
 	WVP = World * XMViewSpace * Projection;
